+++ rocketpy/__init__.py
@@ -18,6 +18,7 @@ from .motors import (
     MassFlowRateBasedTank,
     Motor,
     PointMassMotor,
+    RingClusterMotor,
     SolidMotor,
     SphericalTank,
     Tank,